requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
Faker==4.1.1
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
requests==2.24.0
selene==2.0.0a28
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.2.5
webdriver-manager==3.2.1
zipp==3.1.0
